# Wedding/Wedding.API/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Wedding.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/CardManagement": {
      "get": {
        "tags": [
          "CardManagement"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "filterOn",
            "type": "string"
          },
          {
            "in": "query",
            "name": "filterQuery",
            "type": "string"
          },
          {
            "in": "query",
            "name": "sortBy",
            "type": "string"
          },
          {
            "in": "query",
            "name": "isAscending",
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "pageNumber",
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32",
            "default": 10
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      },
      "post": {
        "tags": [
          "CardManagement"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateCardManagementDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
    "/api/CardManagement/{id}": {
      "get": {
        "tags": [
          "CardManagement"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      },
      "put": {
        "tags": [
          "CardManagement"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateCardManagementDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "CardManagement"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
...
    "/api/Customer": {
      "get": {
        "tags": [
          "Customer"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "filterOn",
            "type": "string"
          },
          {
            "in": "query",
            "name": "filterQuery",
            "type": "string"
          },
          {
            "in": "query",
            "name": "sortBy",
            "type": "string"
          },
          {
            "in": "query",
            "name": "isAscending",
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "pageNumber",
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32",
            "default": 10
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Customer"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateCustomerDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
    "/api/Customer/{customerId}": {
      "get": {
        "tags": [
          "Customer"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "customerId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
    "/api/Customer/activate/{customerId}": {
      "post": {
        "tags": [
          "Customer"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "customerId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
    "/api/Customer/deactivate/{customerId}": {
      "post": {
        "tags": [
          "Customer"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "customerId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
    "/api/Customer/export/{month}/{year}": {
      "post": {
        "tags": [
          "Customer"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "month",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "year",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
    "/api/Customer/download/{fileName}": {
      "get": {
        "tags": [
          "Customer"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "fileName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateCardManagementDTO": {
      "type": "object",
      "properties": {
        "cardId": {
          "format": "uuid",
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "invitationId": {
          "format": "uuid",
          "type": "string"
        },
        "attendStatus": {
          "type": "string"
        },
        "createdTime": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "ResponseDTO": {
      "type": "object",
      "properties": {
        "result": { },
        "isSuccess": {
          "type": "boolean"
        },
        "statusCode": {
          "format": "int32",
          "type": "integer"
        },
        "message": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "UpdateCardManagementDTO": {
      "type": "object",
      "properties": {
        "cardId": {
          "format": "uuid",
          "type": "string"
        },
        "invitationId": {
          "format": "uuid",
          "type": "string"
        },
        "attendStatus": {
          "type": "string"
        },
        "createdTime": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "UpdateCustomerDTO": {
      "required": [
        "address",
        "birthDate",
        "country",
        "customerId",
        "fullName",
        "gender"
      ],
      "type": "object",
      "properties": {
        "customerId": {
          "format": "uuid",
          "type": "string"
        },
        "gender": {
          "minLength": 1,
          "type": "string"
        },
        "fullName": {
          "minLength": 1,
          "type": "string"
        },
        "birthDate": {
          "format": "date-time",
          "type": "string"
        },
        "country": {
          "minLength": 1,
          "type": "string"
        },
        "address": {
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Please enter your token with this format: \"Bearer YOUR_TOKEN\""
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}